src/main/java/com/example/securityex/controller/AuthController.java
package com.example.securityex.controller;

import com.example.securityex.model.User;
import com.example.securityex.security.JwtTokenProvider;
import lombok.AllArgsConstructor;
import lombok.Getter;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class AuthController {

    private AuthenticationManager authenticationManager;

    private JwtTokenProvider jwtTokenProvider;

    public AuthController(final AuthenticationManager authenticationManager, final JwtTokenProvider jwtTokenProvider) {
        this.authenticationManager = authenticationManager;
        this.jwtTokenProvider = jwtTokenProvider;
    }

    @PostMapping("/auth")
    public AuthenticateReponse authenticate(@RequestBody final User user) {

        Authentication authentication = authenticationManager.authenticate(
                new UsernamePasswordAuthenticationToken(
                        user.getEmail(),
                        user.getPassword()
                )
        );

        SecurityContextHolder.getContext().setAuthentication(authentication);

        return new AuthenticateReponse(jwtTokenProvider.generateToken(authentication));
    }

    @Getter
    @AllArgsConstructor
    private class AuthenticateReponse {
        private String token;
    }

}

